+++ CHANGELOG.md
@@ -147,6 +147,7 @@
 - Bug: may corrupt 8bit message source if it is not encoded in UTF-8 (#366)
 - Enh: make debug info optional (#369)
 - import-mbox.py: mailbox defaults to Create if file is not found (#370)
+- Bug: import --dedup does not cater for messages sent to multiple lists (#373)
 
 ## CHANGES in 0.9b: